fintech v0.1.0 Fintech.TransferAgent.Service

This module provides the functions required to comunicate with the database

Functions

back_inter_transference(account, transference, state, int1, int2)

Function in charge of backing the inter transference into database

back_transference(account, transference, state)

Function in charge of backing the transference into database

check_account(state, transference, int1, int2)

Function in charge of checking account from database

check_failure(state, transference, int1, int2)

Function simulating the failure rate

check_transference(state, transference, type, int1, int2)

Function in charge of validating the transference

get_account_with_transferences(account, transference)

Function in charge of getting from database the data required for the update operation

get_account_with_transferences_inter(account, transference, int1, int2)

Function in charge of getting from database the data required for the update operation